requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
unstructured
faiss-cpu
nest_asyncio
python-dotenv
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
